refactor(core): verplaats MarkdownEditor/Toolbar van components/task naar components/editor
De editor is generiek (gebruikt door FormField én de projectomschrijving), dus de
plek onder components/task/ was misleidend. Verplaatst naar een eigen components/editor/-
map; barrel-export, FormField-import en cov-testnamen/-imports meegewijzigd. Util-imports
(../../utils/*) blijven gelijk (zelfde diepte). Geen gedragswijziging; 100% dekking behouden.
