+++ packages/assessment-core/src/components/editor/MarkdownEditor.vue
