+++ packages/assessment-core/src/index.ts
@@ -59,7 +59,7 @@ export { default as TaskSection } from './components/task/TaskSection.vue'
 export { default as TaskGroup } from './components/task/TaskGroup.vue'
 export { default as TaskItem } from './components/task/TaskItem.vue'
 export { default as FormField } from './components/task/FormField.vue'
-export { default as MarkdownEditor } from './components/task/MarkdownEditor.vue'
+export { default as MarkdownEditor } from './components/editor/MarkdownEditor.vue'
 export { default as ImageField } from './components/task/ImageField.vue'
 export { default as ReferenceSuggestions } from './components/ReferenceSuggestions.vue'
 export { default as ActionPointsOverview } from './components/ActionPointsOverview.vue'
